requirements.txt
boto3
numpy==1.24.3
pillow
onnxruntime